+++ modules/FeatureScout/dlg_FeatureScout.h
@@ -138,7 +138,7 @@ private slots:
 	void saveStl();
 	void updateStlProgress(int i);
 	void updateMarProgress(int i);
-	void renderLUTChanges(QSharedPointer<iALookupTable> lut, int colInd);
+	void renderLUTChanges(QSharedPointer<iALookupTable> lut, size_t colInd);
 private:
 	//create labelled output image based on defined classes
 	template <class T> void CreateLabelledOutputMask(iAConnector & con, const QString & fOutPath);
